User service: improve tests
It looks like it’s best to have only one instance of
FirebaseSimpleLogin, with a single auth callback. Otherwise multiple
callbacks of multiple instances get different confusing responses. This
was especially confusing for unregisterUser method. That’s why this
refactoring.

For the user service test, avoid using service methods to verify the
outcome of other service methods because a failure then becomes
ambiguous: was it because the tested method didn’t work, or because the
method the verified the outcome didn’t work?