UserDataService: test: account for trackers
The “authenticated” event on App.userService triggered by UserDataService when
creating and authenticating with the test user is “heared” by UserTracker that
records last login and registration timestamps, and PaymentTracker that checks
if the service is payed. PaymentTracker waits for UserTracker.

When UserDataService logs in, runs its tests, and quickly logs out, and
trackers are left out, can’t access the data they need any more, and throw
errors.