SubscriptionDialog: test: fix “hidden” failure
A failure was shown in the counts but was not shown in the list: done was being
called multiple times. This is because the test was using the live app’s DOM¹
and there was another instance of SubscriptionDialog in the live app that was
listening to the submit button click.

¹ It would have been too big of a setup to build my own DOM dialog. Probably.
Should the object also render it’s DOM? Ideally. Probably. An Angular directive
would. Probably.