Another partial rewrite
This was triggered by the need to send payment overdue notifications,
and ended up as a larger rewrite, one which I am happy with.

Changed modules from nouns to actions. This makes the naming much more
natural and I don’t have to invent nouns like “inquirer”,
“email-sender”, and “payment-overdue-notifier”.

Elimintated maxcomplexity thing from JSHint because it made noise in a
couple of places where it was not appropriate.

Wrote a few tests and then deleted them all because the logic kept
changing three times during this rewrite and I was becoming inconvenient
to update tests.

Moved search history storage to Firebase.

Rethought custom field computation.