Section row preprocessor
The sentence section rows contain the relative link to the PDF in the
first column. That link’s URL has a “case_title” param which is the file
name that will be sent to the browser and is composed of “Dosar-”, the
case #, another dash, and a seemingly random number.

Without a clear reason, that random number chnages and is seen as a data
change by the system, although nothing else changed in the row, neither
in the PDF contents.

This code removes the random suffix by reducing the file name to
“Dosar-#”.