Allow multiple runs per day
We want to try to run searches twice per day.

Also: to survive Heroku daily cycling, the system repeatedly runs the
search after an hour.

So the final criteria is that if the search was run less than an hour
ago, it must be the repeated search, so skip it.

I want to skip the repeated search to keep the seach history useful. If
it’ll be to heavy, I’ll have to trim it, and then it’ll lose value.