streamline serveUser
Also, record the search even if there are no news. This way I will not
have to run the search again on the cover run.

This was “caused” by the intent to write unit tests for this code, found
it hard, and tried to rethink it; worked well. Now it’ll be easier to
write tests.