A different strategy to get new rows
The reason for this change is that the IDs of the published rows are
reset nightly so that the most recent gets the ID of 1.

To get around that, I store the data of relevant columns of the last
fetched row, and the next time, do a search by all the stored columns to
find the new ID of the row. Then, having the ID, I query for all the
rows having the ID less than it.