Only serve the user about once in 18 hours
The intent is to be able to run this program multiple times per day
without actually doing the work multiple times.

I want to schedule it a couple of times during the night because the API
sometimes fails, and in that case I want it to try again later. And if
it went OK the first time, then running it the second time will not
serve the users it served the first time.