Experimenting with a less abusive search
...but still brute-force.

I thought I would download all the records from the last 3 months. But
it looks like there is no way to know for sure when a row was added: the
id and timestamp are reset every night.

What I’m left with is to download all the rows first, and do the
searches locally, and this is what I’m trying to experiment with here.